// src/domain/model/assets.ts
export class AssetModel {
  type: string;
  aircraftType: string;
  registration: string;
  manufacturingDate: Date;
  engineType: string;
  serialNumber: string;
  ipcNumber: string;
  maintainanceProgram: string;
  certification: string;
  deliveryDate: Date;
  totalFlightHours: number;
  totalFlightCycles: number;
  monthlyFlightHours: number;
  monthlyFlightCycles: number;
  owner: string;
  expiryDate: Date;
  camo: string;
  leaseType: string;
  operator: string;
  customer: string;
}